Upgraded project back to Xcode 16.3
The previous commit to downgrade the project was naively made to try opening the project on an older macbook. This did not have the desired effect, so this commit is to undo the previous one instead of rewriting history. Additionally, The config for the test targets has been corrected to point to the correct Info.plist file.